pyproject.toml
[project]
name = "proxypdf"

[tool.isort]
profile = "black"
lines_after_imports = 2
honor_noqa = true

[tool.black]
line-length = 119

[tool.poetry]
name = "proxypdf"
version = "0.1.0"
description = ""
authors = ["None"]

[tool.poetry.dependencies]
python = "~3.9"
reportlab = "~4.0"
Pillow = "~9.4"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
